python-project/requirements.txt
Flask
transformers
deep-translator
google-generativeai
python-dotenv
torch
